leetcode/golang: add implement of leetcode about remove duplicate list ii
