+++ code/leetcode/golang/easy/leetcode/list/intersection.go
@@ -5,7 +5,7 @@ import "golang/common/api/model"
 // issue-link: https://github.com/Alice52/Algorithms/issues/33
 // assume the two list must intersect and all of them have no circle.
 func GetIntersectionNode(headA, headB *model.ListNode) *model.ListNode {
-    //boundary check
+    // boundary check
     if headA == nil || headB == nil {
         return nil
     }
