+++ code/leetcode/golang/medium/leetcode/list/remove_duplicate_sorted_list.go
